Replace AssemblyScanner partials with explicit discovery types
Introduce DomainDiscoveryPipeline as the scan orchestrator and split
responsibilities into ScannedTypeCatalog, DomainGraphBuilder,
GraphRelationshipBuilder, GraphReflectionMapper, EventEmissionScanner,
EventGraphLinker, HandlerIlScanner, and SubTypeDiscovery. AssemblyScanner
remains a thin facade for BoundedContextBuilder. Removes partial-class
splits across AssemblyScanner.*.cs files.
