+++ DomainModeling/Discovery/AssemblyScanner.cs
@@ -1,18 +1,21 @@
 using DomainModeling.Builder;
+using DomainModeling.Graph;
 
 namespace DomainModeling.Discovery;
 
 /// <summary>
 /// Scans assemblies configured in a <see cref="BoundedContextBuilder"/>
-/// and produces a <see cref="Graph.BoundedContextNode"/> with all discovered types
+/// and produces a <see cref="BoundedContextNode"/> with all discovered types
 /// and their relationships.
 /// </summary>
-internal sealed partial class AssemblyScanner
+internal sealed class AssemblyScanner
 {
-    private readonly BoundedContextBuilder _config;
+    private readonly DomainDiscoveryPipeline _pipeline;
 
     public AssemblyScanner(BoundedContextBuilder config)
     {
-        _config = config;
+        _pipeline = new DomainDiscoveryPipeline(config);
     }
+
+    public BoundedContextNode Scan() => _pipeline.Run();
 }
